requirements.txt
pandas
discord.py
python-dotenv
openpyxl